agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai-tools[sqlalchemy]>=1.14.3",
    "crewai[tools]>=1.14.3",
    "fastapi>=0.136.1",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "supabase>=2.29.0",
    "uvicorn[standard]>=0.46.0",
]
